# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Порошок API",
    "description": "REST API для інтернет-магазину побутової хімії (оптово-роздрібна платформа)",
    "version": "1.0.0",
    "contact": {
      "name": "Порошок"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Замовлення поточного користувача",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список замовлень"
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Створити замовлення (checkout)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "contactName": {
                    "type": "string",
                    "minLength": 2
                  },
                  "contactPhone": {
                    "type": "string",
                    "minLength": 10
                  },
                  "contactEmail": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "companyName": {
                    "type": "string"
                  },
                  "edrpou": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 8
                  },
                  "deliveryMethod": {
                    "type": "string",
                    "enum": [
                      "nova_poshta",
                      "ukrposhta",
                      "pickup",
                      "pallet"
                    ]
                  },
                  "deliveryCity": {
                    "type": "string"
                  },
                  "deliveryWarehouseRef": {
                    "type": "string"
                  },
                  "deliveryAddress": {
                    "type": "string"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "enum": [
                      "cod",
                      "bank_transfer",
                      "online",
                      "card_prepay"
                    ]
                  },
                  "comment": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "loyaltyPointsToSpend": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "paymentProvider": {
                    "type": "string",
                    "enum": [
                      "liqpay",
                      "monobank"
                    ]
                  }
                },
                "required": [
                  "contactName",
                  "contactPhone",
                  "contactEmail",
                  "deliveryMethod",
                  "paymentMethod"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Замовлення створено"
          },
          "400": {
            "description": "Невірні дані або порожній кошик"
          }
        }
      }
    },
...
  },
...
}